builder-apps/cortex-agent/requirements.txt
# FastAPI and server
fastapi[standard]>=0.115.8
uvicorn>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.1
